+++ packages/snaps-rpc-methods/src/middleware/index.ts
@@ -1,0 +1,1 @@
+export * from './preinstalled-snaps';
