+++ packages/snaps-rpc-methods/src/middleware/preinstalled-snaps.ts
