--- packages/snaps-rpc-methods/src/preinstalled/index.ts
@@ -1,1 +1,0 @@
-export * from './middleware';
